requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
pydantic>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0
bcrypt>=4.0.1
pytest>=7.3.1
